Adding dynamic calendar function
' PARAMETERS:
' selected_dates_array  - is the output array it will contain dates in MM/DD/YY format
' month_to_use          - this can be MM/YY or MM/DD/YY format as long as it is considered a date it will work. 
' one_date_only         - this is a True/false parameter which will restrict the function to only allow one date to be selected if set to TRUE
' disable_weekends      - this is a True/false parameter which will restrict the selection of weekends if set to TRUE
' start_date & end_date - This will provide a range of dates which cannot be selected. These are to be entered as numbers. For example start_date = 3 and end_date = 14 the days 3 through 14 will be unavailable to select

At this time I haven't gotten any feedback on this working for REVS from other agencies. Since my time at Anoka is coming to a close I wanted to get the function added. I'd uploaded a version of the revs(now probably outdated) working with the calendar to my github so once I get back to github and it's working I can pull that.